Prefer arrow functions
